Ajustando pagamento e alguns atributos.

+++ ProvaAvonaleAPI/Pagamento.cs
@@ -3,7 +3,7 @@
     public class Pagamento
     {
         public int Id { get; set; }
-        public int Valor { get; set;}
+        public double Valor { get; set;}
         public string Estado { get; set; } = string.Empty;
         public Cartao Cartao { get; set; }
     }
